add padding to header,move menu styles, fixe logo,change header name to lowercase
