+++ src/components/header/styles.js
@@ -1,4 +1,4 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 
 export const HeaderContainer = styled.header`
@@ -9,7 +9,8 @@ export const HeaderContainer = styled.header`
 `;
 
 export const Nav = styled.nav`
-  width: calc(100% - 160px);
+  width: 100%;
+  padding: 0 80px;
   margin: 0 auto;
   display: flex;
   justify-content: space-between;
@@ -17,28 +18,6 @@ export const Nav = styled.nav`
 
 export const LogoLink = styled(Link)`
   cursor: pointer;
-`;
-
-export const MenuList = styled.ul`
-  list-style: none;
   display: flex;
   align-items: center;
-`;
-
-export const MenuItem = styled.li`
-  cursor: pointer;
-  :not(:last-child) {
-    margin-right: 26px;
-  }
-`;
-
-export const MenuLink = styled(Link)`
-  color: #636363;
-  text-decoration: none;
-  ${({ theme }) => {
-    return css`
-      font-family: ${theme.fonts.primary};
-      font-size: ${theme.fontSizes.regular};
-    `;
-  }}
 `;
